Extract validators into separate files
