+++ src/validators/index.ts
@@ -1,5 +1,4 @@
 import Vue from 'vue';
-import type { Constructor } from '../types';
 
 export type Validator = (value: unknown) => string | undefined
 export type VuePropValidator = (value: unknown) => boolean
@@ -30,30 +29,10 @@ export function vuePropValidator(userValidator?: Validator, ...typeValidators: V
   };
 }
 
-export const isOneOf = (allowedValues: readonly unknown[]): Validator => value => {
-  if (!allowedValues.includes(value)) {
-    return `value should be one of "${allowedValues.join('", "')}"`;
-  }
-  return undefined;
-};
+// for internal use only
+export { isInstanceOf } from './isInstanceOf';
+export { isInteger } from './isInteger';
+export { isOneOf } from './isOneOf';
+export { isSymbol } from './isSymbol';
 
-export const isInteger: Validator = value => {
-  if (typeof value !== 'number' || !Number.isInteger(value)) {
-    return 'value should be an integer';
-  }
-  return undefined;
-};
-
-export const isSymbol: Validator = value => {
-  if (typeof value !== 'symbol') {
-    return 'value should be a symbol';
-  }
-  return undefined;
-};
-
-export const isInstanceOf = (parent: Constructor): Validator => value => {
-  if (!(value instanceof parent)) {
-    return `value should be an instance of ${parent.name}`;
-  }
-  return undefined;
-};
+// custom validators provided for user convenience are exported from main entrypoint
