+++ src/validators/isInteger.ts
@@ -1,0 +1,8 @@
+import type { Validator } from '.';
+
+export const isInteger: Validator = value => {
+  if (typeof value !== 'number' || !Number.isInteger(value)) {
+    return 'value should be an integer';
+  }
+  return undefined;
+};
