+++ src/validators/isOneOf.ts
@@ -1,0 +1,8 @@
+import type { Validator } from '.';
+
+export const isOneOf = (allowedValues: readonly unknown[]): Validator => value => {
+  if (!allowedValues.includes(value)) {
+    return `value should be one of "${allowedValues.join('", "')}"`;
+  }
+  return undefined;
+};
