+++ src/validators/isSymbol.ts
@@ -1,0 +1,8 @@
+import type { Validator } from '.';
+
+export const isSymbol: Validator = value => {
+  if (typeof value !== 'symbol') {
+    return 'value should be a symbol';
+  }
+  return undefined;
+};
