+++ type-tests/validators/isInteger.type.spec.ts
@@ -1,0 +1,4 @@
+import { expectType } from 'tsd-lite';
+import { isInteger } from '../../src/validators/isInteger';
+
+expectType<string | undefined>(isInteger(undefined));
