+++ type-tests/validators/isOneOf.type.spec.ts
@@ -1,0 +1,4 @@
+import { expectType } from 'tsd-lite';
+import { isOneOf } from '../../src/validators/isOneOf';
+
+expectType<string | undefined>(isOneOf([])(undefined));
