+++ type-tests/validators/isSymbol.type.spec.ts
@@ -1,0 +1,4 @@
+import { expectType } from 'tsd-lite';
+import { isSymbol } from '../../src/validators/isSymbol';
+
+expectType<string | undefined>(isSymbol(undefined));
